refactor: folder structure and dev script
